@@ -11,16 +11,16 @@
         }
     ],
     "require": {
-        "php": ">=5.3.3",
+        "php": "^7.1.3",
         "pagerfanta/pagerfanta": "^1.1.0|^2.0.0",
-        "symfony/framework-bundle": "~2.3|~3.0|~4.0",
-        "symfony/property-access": "~2.3|~3.0|~4.0",
-        "symfony/translation": "~2.3|~3.0|~4.0",
-        "symfony/twig-bundle": "~2.3|~3.0|~4.0"
+        "symfony/framework-bundle": "~4.3|~5.0",
+        "symfony/property-access": "~4.3|~5.0",
+        "symfony/translation": "~4.3|~5.0",
+        "symfony/twig-bundle": "~4.3|~5.0"
     },
     "require-dev": {
-        "symfony/symfony": "~2.3|~3.0|~4.0",
-        "phpunit/phpunit": "~3.7|~4.0|^5.0"
+        "phpunit/phpunit": "~7.0",
+        "symfony/symfony": "4.3"
     },
     "conflict": {
         "twig/twig": "<1.34|>=2.0,<2.4"
